#1 Refactor code to FastAPI standards
