+++ app/services/convert_util.py
@@ -1,12 +1,16 @@
 import re
+from services import convert as doConvert
+
 
 def doCharMap(text, charMap):
     for srcKey, keyVal in charMap.items():
         text = preg_replace(srcKey, keyVal, text)
     return text
 
+
 def mb_strlen(str):
     return len(str)
+
 
 # returns the i-th byte of the multi-byte string str
 def mbCharAt(str, i):
@@ -16,14 +20,17 @@ def mbCharAt(str, i):
     except:
         pass
 
+
 # returns the javascript 'substring' method equivalent
 def subString(string, frm, to):
     # return mb_substr(string, from, to - from)
     return string[frm:to]
 
+
 def preg_replace(srcKey, keyVal, text):
     #srcKey = "@"+srcKey+"@"
     return re.sub(srcKey, keyVal, text)
+
 
 # New Add
 def replace_eKar_after_whitespace(text):
@@ -34,3 +41,18 @@ def replace_eKar_after_whitespace(text):
         else:
             result.append(char)
     return "".join(result)
+
+
+def unicode_to_bijoy(text: str) -> str:
+    """
+    Example function to convert from Unicode to Bijoy using your custom module.
+    """
+    converter = doConvert.Unicode()  # Adjust to match your actual class or function
+    return converter.convertUnicodeToBijoy(text)
+
+
+def bijoy_to_unicode(text: str) -> str:
+    """
+    Placeholder function for the opposite conversion (Bijoy to Unicode).
+    """
+    return text
