+++ app/static/scripts.js
